rag-mindmap-backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
google-generativeai
python-dotenv # For loading .env files 